Do not execute ReactNativeRenderer_prod.js when __DEV__ is true. (#27830)
Summary:

ReactNativeRenderer_prod.js now checks that __DEV__ is false before executing further code, in the same way that ReactNativeRenderer_dev.js checks to see that it is true before executing further code.

This prevents potential mixing of renderers if code tries to load it.
An example demonstrating such a problem is this issue with React Native Debugger:

https://github.com/jhen0409/react-native-debugger/issues/438